ATR-855: fixed unit test for event overrides
